Starting to wrap things up
The night before the final day is here. Ended up being busier than I anticipated this week so I'm starting to wrap things up on this project. Might try to add a little bit more tomorrow but as it stands there's a main menu and a fail state now. A full loop is here. Tomorrow I'm going to make a build and upload it first thing to make sure there are no build errors before trying to add anything else.