That's a wrap for the game jam
Got the project to a playable MVP and submitted on itch for the game jam. Didn't get everything I wanted to get done but I think I did a pretty bang up job for the amount of time I spent on it. Will be revisiting this project idea to flesh it out more in the future.